+++ halo2_gadgets/src/test_circuits/test_utils.rs
@@ -3,7 +3,7 @@
 use std::{env, fs, path::Path};
 
 use group::ff::PrimeField;
-use rand::rngs::OsRng;
+use rand::rng;
 
 use pasta_curves::{pallas, vesta};
 
@@ -42,7 +42,7 @@ impl Proof {
         let pk = plonk::keygen_pk(params, vk.clone(), &circuit).unwrap();
 
         let mut transcript = Blake2bWrite::<_, vesta::Affine, _>::init(vec![]);
-        plonk::create_proof(params, &pk, &[circuit], &[&[]], OsRng, &mut transcript)?;
+        plonk::create_proof(params, &pk, &[circuit], &[&[]], rng(), &mut transcript)?;
         let proof = transcript.finalize();
 
         Ok(Proof(proof))